+++ src/de/langenmaier/u2r3/db/IntersectionOfRelation.java
@@ -36,7 +36,7 @@ public class IntersectionOfRelation extends Relation {
 	}
 
 	@Override
-	public void createDeltaImpl(long id) {
+	public void createDeltaImpl(int id) {
 		try {
 			dropDelta(id);
 			createDeltaStatement.execute("CREATE TABLE " + getDeltaName(id) + "" +
