+++ src/de/langenmaier/u2r3/db/PropertyDomainRelation.java
@@ -48,7 +48,7 @@ public class PropertyDomainRelation extends Relation {
 	}
 
 	@Override
-	public void createDeltaImpl(long id) {
+	public void createDeltaImpl(int id) {
 		try {
 			dropDelta(id);
 			createDeltaStatement.execute("CREATE TABLE " + getDeltaName(id) + " (" +
